+++ src/MemoryCache.php
@@ -25,7 +25,7 @@ use Throwable;
 /**
  * File based memory storage.
  */
-final class Memory implements MemoryInterface
+final class MemoryCache implements MemoryInterface
 {
     // data file extension
     private const EXTENSION = 'php';