+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to `biurad/biurad-caching` will be documented in this file.
 
+# 0.2.1 - 2020-05-04
+- Improved code quality to prevent breaks and better performance
+- Added support for *var_export* caching using the new *MemoryCache* class
+- Added Doctrine's *PhpFileCache* adapter as *memory* cache adapter
+- Renamed *Caching* class to *FastCache* to avoid class conflict
+- Renamed *CachePass* class to *DoctrineCachePass*
+- Removed support on *Database Adapter* for doctine cache
+- Fixed issue with *Memcached* connection on BiuradPHP and Nette Framework
+- Fixed minor bugs faced with caching
+- Updated README.md file
+
 ## 0.1.10 - 2020-03-28
 - Added doctrine cache adapter *Memcached*
 
@@ -21,7 +32,7 @@ All notable changes to `biurad/biurad-caching` will be documented in this file.
 - Added *CacheBridge* class for better performance working with BiuradPHP Framework
 - Added *Connection* class for easy connection to servers
 - Fixed Security issues regarding caching.
-- Updated RREADME.md file
+- Updated README.md file
 
 ## 0.1.5 - 2019-11-21